requirements.txt
python-telegram-bot
ollama
python-dotenv
apscheduler
